+++ src/fnft_nsep.c
@@ -80,20 +80,19 @@ static inline void update_bounding_box_if_auto(const REAL eps_t,
 
 // Main routine.
 INT fnft_nsep(const UINT D, COMPLEX const * const q,
-        REAL const * const T, UINT * const K_ptr,
+        REAL const * const T, REAL const phase_shift, UINT * const K_ptr,
         COMPLEX * const main_spec, UINT * const M_ptr,
         COMPLEX * const aux_spec, REAL * const sheet_indices,
         const INT kappa, fnft_nsep_opts_t * opts_ptr)
 {
     INT ret_code = SUCCESS;
-    UINT i, K1, K2, M1, M2, D_effective;
+    UINT i, K1, K2, M1, M2;
     INT warn_flags[2] = { 0, 0 }; // 0 = no warning about too many points so
     // far, 1st val is for main spec, 2nd for aux
     COMPLEX *q_preprocessed = NULL;
-    COMPLEX m_q = 1.0;
     
     // Check inputs
-    if (D < 2 || D%2 == 0)
+    if (D < 2 || D%2 == 1)
         return E_INVALID_ARGUMENT(D);
     if (q == NULL)
         return E_INVALID_ARGUMENT(q);
@@ -112,28 +111,17 @@ INT fnft_nsep(const UINT D, COMPLEX const * const q,
     if (opts_ptr->filtering != fnft_nsep_filt_NONE && main_spec == NULL && aux_spec != NULL)
         return E_INVALID_ARGUMENT(main_spec. Filtering of the auxiliary spectrum is not possible if the main spectrum is not computed.);
       
-        
-        if (CABS(q[0]) == 0){
-            WARN("First sample of signal q is 0. Not possible to determine if signal is quasi-periodic. Continuing assuming signal is periodic. Computed spectrum maybe incorrect.");
-        }else
-            m_q = q[D-1]/q[0];
-        
-        if (FABS(1-CABS(m_q)) > 1e-2)
-            WARN("Signal q may not be (quasi-)periodic. abs(1-abs(q[D-1]/q[0]))>1e-2.");
-        
-        REAL Lam_shift = CARG(m_q)/(-2*(T[1] - T[0]));
-        const REAL eps_t = (T[1] - T[0])/(D - 1);
-        D_effective = D-1;
-        q_preprocessed = malloc(D_effective * sizeof(COMPLEX));
+        REAL Lam_shift = CARG(CEXP(I*phase_shift))/(-2*(T[1] - T[0]));
+        const REAL eps_t = (T[1] - T[0])/D;
+        q_preprocessed = malloc(D * sizeof(COMPLEX));
         if (q_preprocessed == NULL) {
             ret_code = E_NOMEM;
             goto leave_fun;
         }
         
         // Removing phase rotation along q
-        for (i=0; i < D_effective; i++)
+        for (i=0; i < D; i++)
             q_preprocessed[i] = q[i]*CEXP(2*I*Lam_shift*(T[0]+eps_t*i));
-        // Last sample is dropped as it is the beginning of next period
         
     
     // Bounding box needs to be shifted to ensure it works properly for
@@ -154,13 +142,13 @@ INT fnft_nsep(const UINT D, COMPLEX const * const q,
                 // Compute non-real points in the spectra via subsample & refine
                 
                 if (kappa == +1) {
-                    ret_code = subsample_and_refine(D_effective, q_preprocessed, T, &K1, main_spec,
+                    ret_code = subsample_and_refine(D, q_preprocessed, T, &K1, main_spec,
                             &M1, aux_spec, sheet_indices, kappa, opts_ptr,
                             1 /*skip_real_flag*/, warn_flags);
                     CHECK_RETCODE(ret_code, leave_fun);
                 } else { // no non-real main spec in the defocusing case, pass NULL
                     K1 = 0;
-                    ret_code = subsample_and_refine(D_effective, q_preprocessed, T, &K1, NULL,
+                    ret_code = subsample_and_refine(D, q_preprocessed, T, &K1, NULL,
                             &M1, aux_spec, sheet_indices, kappa, opts_ptr,
                             1 /*skip_real_flag*/, warn_flags);
                     CHECK_RETCODE(ret_code, leave_fun);
@@ -175,7 +163,7 @@ INT fnft_nsep(const UINT D, COMPLEX const * const q,
                 
                 // Compute real points in the spectra via gridsearch
                 
-                ret_code = gridsearch(D_effective, q_preprocessed, T, &K2, main_spec+K1,
+                ret_code = gridsearch(D, q_preprocessed, T, &K2, main_spec+K1,
                         &M2, aux_spec+M1, sheet_indices, kappa, opts_ptr, warn_flags);
                 CHECK_RETCODE(ret_code, leave_fun);
                 
@@ -188,7 +176,7 @@ INT fnft_nsep(const UINT D, COMPLEX const * const q,
                 
             case fnft_nsep_loc_SUBSAMPLE_AND_REFINE:
                 
-                ret_code = subsample_and_refine(D_effective, q_preprocessed, T, K_ptr, main_spec,
+                ret_code = subsample_and_refine(D, q_preprocessed, T, K_ptr, main_spec,
                         M_ptr, aux_spec, sheet_indices, kappa, opts_ptr,
                         0/*skip_real_flag*/, warn_flags);
                 CHECK_RETCODE(ret_code, leave_fun);
@@ -196,7 +184,7 @@ INT fnft_nsep(const UINT D, COMPLEX const * const q,
                 
             case fnft_nsep_loc_GRIDSEARCH:
                 
-                ret_code = gridsearch(D_effective, q_preprocessed, T, K_ptr, main_spec,
+                ret_code = gridsearch(D, q_preprocessed, T, K_ptr, main_spec,
                         M_ptr, aux_spec, sheet_indices, kappa, opts_ptr, warn_flags);
                 CHECK_RETCODE(ret_code, leave_fun);
                 break;