+++ test/fnft_nsep/fnft_nsep_test_numerical_focusing_3.c
@@ -321,7 +321,9 @@ INT fnft_nsep_test_numerical_focusing_3()
     opts.bounding_box[1] = 3;
     opts.bounding_box[2] = -10;
     opts.bounding_box[3] = 10;
-    ret_code = fnft_nsep(D, q, T, &K, mainspec, &M, auxspec, NULL, +1, &opts);
+    
+    REAL phase_shift = CARG(q[D-1]/q[0]);
+    ret_code = fnft_nsep(D-1, q, T, phase_shift, &K, mainspec, &M, auxspec, NULL, +1, &opts);
     CHECK_RETCODE(ret_code, leave_fun);
 
     // Check that the main and auxiliary spectrum are correct